refactor sqlite code
