+++ src/sqlite.js
@@ -1,0 +1,22 @@
+const log = require("log-less-fancy")();
+const sqlite3 = require("sqlite3"); //.verbose();
+
+function safe(arg) {
+  return arg.replace(/[^0-9a-z]/gi, "");
+}
+
+function dball(file, sql, args) {
+  log.info("Open " + file);
+  return new Promise((resolve, reject) => {
+    const db = new sqlite3.Database(file, sqlite3.OPEN_READONLY, err => {
+      if (err) return reject(err);
+      db.all(sql, args, (err, records) => {
+        db.close();
+        if (err) return reject(err);
+        resolve(records);
+      });
+    });
+  });
+}
+
+module.exports = { safe, dball };
