+++ src/sqliteReader.js
@@ -1,13 +1,9 @@
 const log = require("log-less-fancy")();
-const sqlite3 = require("sqlite3"); //.verbose();
-
-function safe(arg) {
-  return arg.replace(/[^0-9a-z]/gi, "");
-}
+const { safe, dball } = require("./sqlite");
 
 async function read(file, table, key) {
   log.info(`Read key ${key} from ${table} in file ${file}`);
-  const rows = await execute(
+  const rows = await dball(
     file,
     "SELECT verdi from " + safe(table) + " WHERE kode=?",
     [key]
@@ -18,36 +14,19 @@ async function read(file, table, key) {
 }
 
 async function listRows(file, table) {
-  return await execute(file, "SELECT kode FROM " + safe(table) + " LIMIT 10");
+  return await dball(file, "SELECT kode FROM " + safe(table) + " LIMIT 100");
 }
 
 async function listTables(file, filter) {
-  return await execute(
+  return await dball(
     file,
     "SELECT name FROM sqlite_master WHERE type IN ('table','view') AND name NOT LIKE 'sqlite_%' ORDER BY 1"
   );
 }
 
 async function getColumns(file, table) {
-  const records = await execute(
-    file,
-    "PRAGMA table_info('" + safe(table) + "')"
-  );
+  const records = await dball(file, "PRAGMA table_info('" + safe(table) + "')");
   return records.map(rec => rec.name);
-}
-
-function execute(file, sql, args) {
-  log.info("Open " + file);
-  return new Promise((resolve, reject) => {
-    const db = new sqlite3.Database(file, sqlite3.OPEN_READONLY, err => {
-      if (err) return reject(err);
-      db.all(sql, args, (err, records) => {
-        db.close();
-        if (err) return reject(err);
-        resolve(records);
-      });
-    });
-  });
 }
 
 module.exports = { read, listTables, listRows, getColumns };
